Move functionality from fuzzclaim
