+++ internal/fuzzbinder/fuzzclaim.go
@@ -26,7 +26,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"time"
 
 	"github.com/project-oak/transparent-release/pkg/amber"
 	"github.com/project-oak/transparent-release/pkg/intoto"
@@ -178,139 +177,5 @@ func parseFuzzClaimBytes(statementBytes []byte) (*intoto.Statement, error) {
 		return nil, fmt.Errorf("could not validate the parsed fuzzing claim: %v", err)
 	}
 
-	return &statement, nil
-}
-
-// TODO(#171): Split generateFuzzClaimSpec into smaller functions.
-// generateFuzzClaimSpec generates a fuzzing claim specification using the
-// fuzzing reports of OSS-Fuzz.
-func generateFuzzClaimSpec(revisionDigest intoto.DigestSet, fuzzParameters *FuzzParameters, fuzzTargets []string) (*FuzzClaimSpec, error) {
-	var projectCrashes Crash
-	var projectFuzzEffort FuzzEffort
-	fuzzersCrashes := make(map[string]*Crash)
-	fuzzersFuzzEffort := make(map[string]*FuzzEffort)
-	fuzzersCoverage := make(map[string]*Coverage)
-	//Get fuzzing statistics.
-	for _, fuzzTarget := range fuzzTargets {
-		coverage, err := GetCoverage(fuzzParameters, fuzzTarget, "perTarget")
-		if err != nil {
-			return nil, fmt.Errorf(
-				"could not get %s coverage to generate the fuzzing ClaimSpec: %v", fuzzTarget, err)
-		}
-		fuzzEffort, err := GetFuzzEffort(revisionDigest, fuzzParameters, fuzzTarget)
-		if err != nil {
-			return nil, fmt.Errorf(
-				"could not get %s fuzzing efforts to generate the fuzzing ClaimSpec: %v", fuzzTarget, err)
-		}
-		crash, err := GetCrashes(revisionDigest, fuzzParameters, fuzzTarget)
-		if err != nil {
-			return nil, fmt.Errorf(
-				"could not get %s crashes to generate the fuzzing ClaimSpec: %v", fuzzTarget, err)
-		}
-
-		fuzzersCrashes[fuzzTarget] = crash
-		fuzzersFuzzEffort[fuzzTarget] = fuzzEffort
-		fuzzersCoverage[fuzzTarget] = coverage
-
-		projectCrashes.Detected = projectCrashes.Detected || crash.Detected
-		projectFuzzEffort.FuzzTimeSeconds += fuzzEffort.FuzzTimeSeconds
-		projectFuzzEffort.NumberFuzzTests += fuzzEffort.NumberFuzzTests
-	}
-	projectCoverage, err := GetCoverage(fuzzParameters, "", "perProject")
-	if err != nil {
-		return nil, fmt.Errorf(
-			"could not get the project coverage to generate the fuzzing ClaimSpec: %v", err)
-	}
-	// Generate fuzzing claim specification.
-	perProject := &FuzzStats{
-		BranchCoverage:  projectCoverage.BranchCoverage,
-		LineCoverage:    projectCoverage.LineCoverage,
-		DetectedCrashes: projectCrashes.Detected,
-		FuzzTimeSeconds: projectFuzzEffort.FuzzTimeSeconds,
-		NumberFuzzTests: projectFuzzEffort.NumberFuzzTests,
-	}
-	perTarget := make([]FuzzSpecPerTarget, 0, len(fuzzTargets))
-	for _, fuzzTagret := range fuzzTargets {
-		targetStats := FuzzStats{
-			BranchCoverage:  fuzzersCoverage[fuzzTagret].BranchCoverage,
-			LineCoverage:    fuzzersCoverage[fuzzTagret].LineCoverage,
-			DetectedCrashes: fuzzersCrashes[fuzzTagret].Detected,
-			FuzzTimeSeconds: fuzzersFuzzEffort[fuzzTagret].FuzzTimeSeconds,
-			NumberFuzzTests: fuzzersFuzzEffort[fuzzTagret].NumberFuzzTests,
-		}
-		targetSpec := FuzzSpecPerTarget{
-			Name:      fuzzTagret,
-			Path:      fmt.Sprintf("%s/fuzz/fuzz_targets/%s.rs", fuzzParameters.ProjectName, fuzzTagret),
-			FuzzStats: &targetStats,
-		}
-		perTarget = append(perTarget, targetSpec)
-	}
-	fuzzClaimSpec := FuzzClaimSpec{
-		PerTarget:  perTarget,
-		PerProject: perProject,
-	}
-	return &fuzzClaimSpec, nil
-}
-
-// GenerateFuzzClaim generates a fuzzing claim (an instance of intoto.Statement,
-// with AmberClaimV1 as the PredicateType and FuzzClaimV1 as the ClaimType) using the
-// fuzzing reports of OSS-Fuzz and ClusterFuzz.
-func GenerateFuzzClaim(fuzzParameters *FuzzParameters) (*intoto.Statement, error) {
-	revisionDigest, err := GetCoverageRevision(fuzzParameters)
-	if err != nil {
-		return nil, fmt.Errorf(
-			"could not get the revision digest to generate the fuzzing claim: %v", err)
-	}
-	fuzzTargets, err := GetFuzzTargets(fuzzParameters)
-	if err != nil {
-		return nil, fmt.Errorf(
-			"could not get the fuzzing targets to generate the fuzzing claim: %v", err)
-	}
-	fuzzClaimSpec, err := generateFuzzClaimSpec(revisionDigest, fuzzParameters, fuzzTargets)
-	if err != nil {
-		return nil, fmt.Errorf(
-			"could not get the fuzzing ClaimSpec to generate the fuzzing claim: %v", err)
-	}
-	evidences, err := GetEvidences(fuzzParameters, fuzzTargets)
-	if err != nil {
-		return nil, fmt.Errorf(
-			"could not get evidences to generate the fuzzing claim: %v", err)
-	}
-	currentTime := time.Now()
-	tomorrow := time.Now().AddDate(0, 0, 1)
-	// TODO(#173): Add validity duration as an input parameter.
-	endValidity := time.Now().AddDate(0, 0, 7)
-	validity := amber.ClaimValidity{
-		NotBefore: &tomorrow,
-		NotAfter:  &endValidity,
-	}
-	// Generate Amber predicate
-	predicate := amber.ClaimPredicate{
-		ClaimType: FuzzClaimV1,
-		ClaimSpec: *fuzzClaimSpec,
-		IssuedOn:  &currentTime,
-		Validity:  &validity,
-		Evidence:  evidences,
-	}
-	// Generate intoto statement
-	subject := intoto.Subject{
-		Name:   fuzzParameters.ProjectGitRepo,
-		Digest: revisionDigest,
-	}
-	statementHeader := intoto.StatementHeader{
-		Type:          intoto.StatementInTotoV01,
-		PredicateType: amber.AmberClaimV1,
-		Subject:       []intoto.Subject{subject},
-	}
-	statement := intoto.Statement{
-		StatementHeader: statementHeader,
-		Predicate:       predicate,
-	}
-	validFuzzPredicate, err := ValidateFuzzClaim(statement)
-	if err != nil {
-		return nil, fmt.Errorf(
-			"could not validate the generated fuzzing claim: %v", err)
-	}
-	statement.Predicate = validFuzzPredicate
 	return &statement, nil
 }
